Modify Space class to include a 3D grid of voxels
A new Voxel class has been added. The Space class has been modified to
include a storage vector of voxels. This vector is initialised upon
construction of the space object. The space constructor now requires a
grid step and a (octree) depth. The old constructor has yet to be
removed.