Add function to sum voxels and calculate volume
A member function has been added in the space class, that sums all
voxels in the grid that are of a specified type. This utilises a member
function in the Voxel class. Whenever a mixed voxel is encountered, the
subvoxels stored in the Voxel object are assessed. When tallying up the
voxels, the result is given in units of bottom level voxels so that the
integer result only has to be multiplied with the unit voxel volume to
receive the total volume.