Fix issue where octree depth changes result
An issue in the 'placeAtomsInGrid' method of the Space class was
identified. Previously the position of the top level voxels was
calculated incorrectly. Starting from the space origin, half the top
level voxel side length has to be added to each coordinate. Before this
fix, half the bottom level voxel side length was used instead. Now, the
depth parameter only changes the runtime and not the result.