Add tree class and node struct for atom tree
The previously added AtomTree class has been modified. The AtomTree
class now points to the binary tree's root and manipulation of the tree
should be done through this class. The tree's root is represesented by
instances of a new AtomNode struct, containing pointers to the atom, and
two AtomNode children. The tree can be constructed and printed in the
console.