Update base_init.cpp
See related issue report (pasted below)
By the way, I saw that there are other instances of SetSizer for the panels. If you wanted them to behave in some specific way based on SetSize, you might want to fix them too.
"When starting the program, the window appears with a default size showing only the browse panel and the parameters panel but not the send and receive panel.

After testing, there are two causes to that in base_init.cpp :
1) The SetSizer() function that is meant to resize the MainWin to the sum of heights of its child frames should be changed to SetSizerAndFit().

2) The child panels don't actually have sizes since they were created with wxDefaultSize which is equal to -1,-1. Therefore, the SetSizer function can only lead to -3 by adding all vertical heights.

I will submit fix to address this issue. If you have better ideas, feel free to reject the proposition."