Replace function that determines voxel type
The determineType method in the Voxel class has been replaced with a
method of the same name. The new method uses a binary tree of atoms
rather than a list of atoms in order to determine the voxel type. This
change massively improves performance, finishing a computation that
previously took 30s in only 1s.