Rework readAtomsFromFile method in Model class
As part of this method, lines from the input file need to be analysed to
determine whether they hold valid atom information, i.e., are of format
<Element Symbol; x-coordinate; y-coordinate; z-coordinate>. Previous to
this change, this assessment was only based on the first character of
each of these properties. For instance, if the x-coodinate was contained
in a substring "3.5", only the first character '3' would be assessed and
the whole string would be deemed numerical. Now, the whole substring is
taken into account.