Fix issue #16, evaluate input coordinate correctly
The isAtomLine() function in Model class has been reworked. With this
new change, only lines in the input file will be considered, in which
the x-, y-, and z-coordinates are purely numerical. Previously, a
coordinate such as "3x15" would have been valid, and interpreted as "3".
This is no longer the case.