Move GUI control to controller in load event
When the user presses the "Load" button, the event function
OnLoadFiles() of the MainFrame class is called. Previously this function
handled some GUI elements without supervision of the Controller class.
Now, GUI control has been entirely exported to loadInputFiles() function
of the Controller class.

This function
- disables GUI elements when the button is pressed
- imports both radius and atom files
- calls MainFrame::generateAtomList() to initialise the Grid control
- reactivates GUI elements

Furthermore, this function does not create an additional instance of the
Model class.