Remove unneeded code and follow formula convention
The commit prior to this added a function that made another function
obsolete. Parts of the code that used this unneeded function have been
rewritten and the function has been removed. Furthermore, the output of
the chemical formula has been reworked to follow common convention of
listing carbon first, hydrogen second, and all other atoms
alphabetically.