Move maps to private and refactor controller code
Various changes have been made to facilitate moving radius and atom
number maps in the Model class from public to private access. As part of
this change, code in a controller method was changed, to have take a map
generated by the view-side and pass it onto the model-side with the
controller as mediator.