Add code to save import file paths and write times
Upon import the file paths and write times are saved by the Model class.
In future changes, it will be possible to compare the input file meta
data with the meta data of the previously imported files, to determine
whether a new import is necessary.

Comparisons have already been added to the code and tested, they do not
affect the software behaviour yet.