Allow start button to import files when needed
When a calculation is started, the currently input atom file and radius
file are checked for changes in order to determine whether another
import is needed. If the file paths have remained the same AND the files
have not been rewritten since the last import, the import is skipped.
Otherwise, the import is started, by essentially "pressing" the load
button.