Remove second-hand reference
The vdw radii values were obtained from the scientific article provided as reference ( S. Alvarez, Dalton Trans. 2013, 42, 8617-8636. https://doi.org/10.1039/C3DT50599E ).

In the previous version of this file, I also provided a secondary reference (WebElements, https://www.webelements.com ) claiming to provide vdw radii values based on the article cited above. However, after checking the WebElements website, it turns out that some values were wrongly copied leading to erroneous vdw radii for elements with Z > 60.

The main purpose for providing the WebElements as a secondary reference was its free access and simple use. However, since the data it contains is wrong, there is no reason to keep it as a secondary reference.

I contacted the website owner months ago but the values are not fixed yet (and might never be since it would invalidate all the merchandise sold by this website, but this is pure personal speculation).