Add map to store default states of gui controls
When GUI elements are disabled using the enableGuiElements() function,
all interactable GUI elements are disables. When they are reenabled,
only those elements are enabled, that have their default state set to
true.

The default states of all interactable GUI controls is stored in a map
inside the MainFrame class in base.h. The default states can be obtained
by using the GUI element object as key.

The map is initialised in base_init.cpp. A function is run when the
software is started, that initialised the map. The default accessibility
of all GUI elements has to be set manually in this function.

Afterwards, when the accessibility of a widget changes, its default
state can be set with setDefaultState(wxWindow*, bool). A set of toggle
functions have been added to base_event.cpp, that act as wrapper
functions for setDefaultState().