Add method to link atoms to neighbours
Added a set of methods in AtomTree and Model classes to link atoms to
adjacent atoms. Adjacent atoms are defined as atoms whose shells are a
set distance apart. The shell radius is given by the atom radius.

The linking is done through pointers in the Atom structures. Adjacent
atom pointers are stored in a vector.