Add unit cell analysis function
The main addition is the function to analyze crystal unit cells.
I also added extra parameters in the GUI for future features (such as probe mode and probes radii).
I made a new source file for functions related to file writing (report, structures, surface maps). Note that I don't use the makefile in the repository so I did not update the makefile.