Add function isPointBetween
Determine whether a point lies in the space between two other points.
Currently, this function requires that the vector pointing to the point
and the vector pointing from the lower to the higher bound originiate
from the same point. This may be changed if needed.

Used to quickly filter voxels that are out of reach for an access
evaluation.