Change radiusfile import to use custom data struct
A new import function has been added to the radius file import routine.
This function reads the selected radius file and builds two unordered
maps. These maps are later used for looking up the radius and the atomic
number respectively of a given atom (indexed by its element symbol).

The function uses a new custom struct called RadiusFileBundle to return
both maps in an easily managable way. As part of the import routine,
these maps are saved as attributes of the Model class.

These changes have been made in order to prepare the implementation of
command line features, i.e., the ability to run a calculation from the
command line. In order to facilitate this, the GUI must be bypassed.

Another function has been added to the Model class, that allows the
import of a radius map directly from the file path. This bypasses the
GUI.