Return volumes for multiples voxel types
Previously the Space::calcVolume() function was only able to return
one double value. This was used to retrieve the vdW Volume. The
function now returns a map with volumes for all types that are
specified in the function. At this point this is only used for the vdW
Volume and the probe excluded volume but can be expanded to any other
type.