Refactor AtomNode and AtomTree to avoid pointers
AtomNode was changed so that instead of containing a pointer to an Atom
object, it now holds an index with which to access the Atom. The list to
which the index belongs is saved as static member in the AtomNode class.

AtomNode has been changed from a struct to a class in order to increase
code safety. Some methods and members have been moved to private.

The AtomTree class as well as code in the Voxel class have been changed
in order to work as previously with the new changes.