Improve function getting atom list from tree
The function that assembles a list of atoms a certain distance from a
given voxel was significantly optimised. Previously each time the
function was called, a new std::vector was created. The list ws
assembled recursively, which required heavy use of the "insert"
function. The use of "insert" slowed down list creation significantly.