Increase size of cell to include unassigned voxels
Previously, it was possible that the voxel grid was too small to include
all voxels that remain unassigned after the first type evaluation
(i.e. potential shell type voxels). This leads to problems when the type
of these voxels need to be determined. The types 'shell' and 'excluded'
are discriminated by the distance of a voxel to the closest 'core' type
voxel. However, if the nearest 'core' type voxel is outside the grid,
the voxel type cannot be determined.

With this change, the probe1 radius is taken into account when
determining the size of the cell/grid. As a consequence, all voxels
bordering the edge of the cell are of type 'core'.