Add split-function to 2nd round of type assignment
Analogously to the 1st round of the type assignment algorithm, a
function has been added for the 2nd round that is able to split a
childless voxel into 8 subvoxels, i.e. add 8 children. This will
be used to exploit the octree data structure and accelerate the
2nd round of the algorithm.

The splitting does not affect the volumes computed.