Add auxillary struct to manage neighbour searches
The SearchIndex struct has been added, in order to create a useful
static object for the Voxel class. An instance of SearchIndex is
initialised before the calculation begins and contains all information
needed for the neighbour search conducted in the second round of the
type assignment algorithm.

This addition solves an issue where the upper search limit could end up
larger than the size of the search index list, resulting in an access
out of bounds and hence a segmentation fault. This is solved with the
struct, by determining the largest possible search limit before the
calculation and adjusting the size of the search index list
appropriately.