fix rounding error
An implicit type cast caused rounding errors in rare cases in such a way that the number of voxels for a unit cell did not match the real values.
With this explicit type cast, the end index and mod index should always be correctly calculated in limit cases (i.e. +1 for end index and 0 for mod index or +0 for end index and 1 for mod index).