Add way to inform user about errors though pop-ups
Added the infrastructure for the controller to open an error dialog or
print an appropriate error message to the console. The controller may
receive a predefined error code that is evaluated by the controller
using a static lookup table. The resulting error message is then either
send to the GUI, causing a dialog box to pop up, or printed to the
console if the app has been operated from the command line.

One error message has been added, that is used when the maximum number
of cavities has been exceeded.