Make abort button able to end running calculation
In order for the abort button to end a running calculation, the
worker thread has to periodically check whether it has received the
signal to stop the calculation. This has been implemented for part of
the volume algorithm using a custom exception. Additionally a
wxMessageQueue was added in order to communicate the abort instruction
to the worker thread.