Set rules for valid element symbols during import
Following the discussion in #17, the function strToValidSymbol() inside
the Model class has been adjusted. Valid element symbols are a string of
alphabetic characters including underscore with the first letter in
uppercase and all subsequent letters in lowercase. Any non-alphabetic
characters afterwards are ignored. If the symbol begins with a
non-alphabetic character then the function returns an empty string.

The function is used during structure file and radius file imports and a
error message is displayed, in case of an invalid symbol.

Resolves: #17