Catch exception for unused PDB file properties
In the AtomLinePDB struct inside model_filereading.cpp an error is
thrown when std::stoi or std::stod fail to convert the inputs. However,
most members inside the class are not used and only included in case
they become useful one day. Furthermore, some of the properties inside
pdb ATOM and HETATM lines are optional, so that there should be no
exception thrown if the extracted string is empty or contains only
whitespaces.

This commit fixes the import issue by catching all invalid_argument
exceptions from std::stoi and std::stod for unused properties, and
setting those values to 0.

Resolves: #84