Limit scope of function to improve clarity
The function "generateChemicalFormula()" that was previously a method of
the Model class was rewritten. It is now slightly shorter and no longer
part of Model. These changes aim to improve readability. Previously the
function took no arguments, so the input and output was obscured. With
this change, the function has a return value and input arguments. Hence,
when the function is called in generateVolumeData(), it is clear where
the output is stored.

The scope of the function has also been limited, because it was only
called once. Therefore, it is not necessary to have the function be
available to the entire Model class. The function declaration has been
placed before the definition of generateVolumeData().