Restructure functions that call volume calculation
Changes have been made to the chain of functions that were called in
order to begin the type assignment algorithm. Previously, there was a
long chain of functions that were called in sequence.

runCalculation() calls generateData() calls generateVolumeData() calls
calcVolume(), and only then is the function called that begins the type
assignment. This is somewhat confusing. The contents of calcVolume()
have been moved to generateVolumeData(). The part of
generateVolumeData() that was tasked with making preparations for the
calculation has been placed in a separate function.