Rework command line output of cavity list
Rework introduces three new structs for storing the whole table to be
displayed. These structs are meant to be used to manage both command
line and GUI outputs.

Columns of the output can now be hidden if they aren't interesting, for
instance the surface columns or the cavity type column.