Change how table formatting is done in report
In the report file a table is used to display the cavity data. This
tables used to use tab characters to format the field. However, this
solution is not portable since tab length is variable. Instead, the code
section that handles the table was rewritten to use the GridData struct
and the field function in order to produce a portable table format.