Make access to private member indirect
Previously, Model::convertCIfAtomList() directly accessed the private
member _cart_matrix. Here the direct access changes to an indirect
access using _cart_matrix as an input argument. The purpose of this
change is to reduce the use of private members as global variables in
order to make the code more readable.

As part of this change, the type of _cart_matrix was changed from an
array to a std::array, so it can be passed as an argument.