Remove global variables and potential bug fix
Added a return value containing the crystallographic import data to
Model::convertCifSymmetryElements, in order to reduce the use of private
members as global variables, and therefore make the function more
readable.

Is appears like, previous to this change, calculations that were
conducted using CIF files in the unit cell mode produced an incorrect
result.