Move CIF import function to FileMngr and add Cryst
Moved the CIF import function to the new import manager. Since the
import function used Model::orthogonalizeUnitCell(), that function was
moved to a new namespace itself. This is to avoid the import manager
having to know about Model and to keep the code separate. This new
namespace is Cryst and should contain functions that are used for
processing unit cells. There are still functions within Model that would
need to be refactored and could then be incorporated into Cryst.