Fix crash when loading an invalid elements file
Previously, the application would crash, if the user tried to load a
structure file, while the path to the elements file was invalid. This
was because when an invalid elements file was read, the map containing
the atom radii was not initialised. This was changed, so that the map is
initialised as an empty map, even when the import was unsuccessful.