Append charge to element symbol in PDB file import
The function strToValidSymbol() now accepts an additional optional
argument for the charge of the atom. The charge, if it is not equal to
0, is appended to the atom symbol, followed by the sign. I.e., if the
imported element has the symbol "Ca" and a charge of -2, the resulting
output from the function will be "Ca2+". This is in accordance with
common convention. Currently, this is only relevant to PDB files.