Change default radius for atoms w/ unknown charges
Added a function that strips the charge from a atom symbol. To be
precise, the function finds the first non-letter character and removes
all subsequent characters. This function is now used by the radius
look-up function if no radius could be found within the elements file.
This way, a charged atom for which no radius could be found, defaults to
the van-der-Waals radius of the element.