Add unit test in CMake
Added the infrastructure for unit tests using ctest through CMake.
Running unit tests requires building a library from the source code.
Errors occur when wxWidgets is included in the library through base.h,
so the library source files also cannot include base.h or any other file
that includes base.h. This is a significant limitation with the current
code and may limit what kind of unit tests can be written. It may
inspire future code to be more modular.

Since importmanager.cpp includes controller.h, it was necessary to add a
custom flag that removes all uses of the Ctrl class. This is ugly but
should be seen as a temporary fix.