Add function to evaluate atom occurences
Added function atomCount, that is capable of evaluating the number of
each element on the fly from a list of atom data. This is fix an issue,
where the incorrect molar mass and molecular formular were constructed,
when elements were ex- or included in calculations.

It would be good to eventually remove _atom_count in Model, because the
information within is redundant.