Fix issue with incorrect molar mass in report
_atom_count and _unit_cell_atom_count were removed from Model. There was
a bug where an incorrect chemical formula and molar mass were provided,
if atoms were excluded and then included again. This was due to the fact
that the contents of the _atom_count containers were deleted and not
reset when elements were included.

Since the information inside the _atom_count containers is also
contained in the _atom_coordinates containers, it was redundant to keep
them around. The function atomCount has been added that allows
evaluating the atom counts on the fly.

This caused another issue. Due to the way that unit cell mode works, the
list of atoms considered for the calculation is not identical to the
list of atoms considered for the molar mass/chemical formula. Changes
have been made to make sure that the correct lists are used.