Correct defaulting behaviour for atomic weights
Previously, if charged atoms were not listed in the elements file then
they would not be attributed a molecular weight, even if the base atom
was listed. For instance, an ion such as Cu2- should be attributed the
atomic weight of copper. With this change, this behaviour has been
introduced.