Add link-time optimisation to increase performance
After moving the method definitions for the Atom struct from its .h file
to a new .cpp file, the program execution slowed down slightly. It
turned out that the compiler could not perform the same optimisation
when definitions and declarations were separate. To allow this
optimisation nonetheless required enabling link-time optimisation. With
this change, program execution is around 30% faster for the release
build.