Add grace period where new files cannot be deleted
The function that deletes old input and output files now has an optional
argument to determine the minimum age that files need to have before
they can be considered for deletion. If the files are younger than the
grace period they will not be deleted. This is to avoid conflicts when
two users are using the app at the same time.