Make sure files are deleted when space runs out
Adjusts the space allowance and grace period for files on the web
server. Adds a new argument to the clean up function that sets a max
amount of storage dedicated to a directory. If the directory exceeds the
storage limits then files will be deleted regardless of the grace period.