Install MoloVol directly to server
Instead of building the executable in a build directory and running it
directly from there, it is much safer to build a deb file for
installation and use it to install MoloVol to the system. This makes
MoloVol accessible directly through the command "molovol" by adding the
executable to the PATH. Hence launch_headless.sh can directly call the
command instead of running the executable from the build folder.

The instructions for installation have been added to the Dockerfile.

The benefit of this is that resource files are installed on a system
directory instead of in a local directory. This makes the setup less
susceptible to subtle changes, such as changing the working directory,
binary directory, or resource directory.

This change required an adjustment in app.py, so that the system
elements file is accessed instead of the local elements file.