Remove static members and methods from AtomTree
Previously, the atom data inside of the AtomTree class was stored in a
static member vector. The problem with this approach is that this
permits only one unique AtomTree instance at a time. To avoid future
bugs, this change removes the static members and stores them instead
inside the AtomTree object.