Add changes to allow compilation on Windows
These changes make it possible to compile MoloVol on Windows using Cmake and VS Code. Importantly, the Runtime Library is now linked dynamically, as is the default and wxWidgets has to be built the same way. This is necessary as VTK compiles with dynamically linked Runtime Library. It also seems to be the recommended setting.