Modify library paths
Previously, all Windows builds dynamically linked runtime libraries, resulting in executables that need to be shipped with certain DLLs. Now, the library paths have been adjusted to link to specific build directories of statically linked libraries (VTK and wxWidgets).

The problem is still, that the library paths are all hard-coded. Ideally, the library path should be set via a cmake variable or automatically found. This will be tackled at a later time.